+++ app/src/main/res/layout/activity_details.xml
@@ -1,100 +1,6 @@
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
+<fragment xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools" android:id="@+id/fragment"
+    android:name="com.example.snehaanand.moviesapp.DetailsActivityFragment"
+    tools:layout="@layout/fragment_main" android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:orientation="vertical"
-    android:paddingBottom="@dimen/activity_vertical_margin"
-    android:paddingLeft="@dimen/activity_horizontal_margin"
-    android:paddingRight="@dimen/activity_horizontal_margin"
-    android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.example.snehaanand.moviesapp.DetailsActivity">
-
-    <ScrollView
-        android:id="@+id/movie_details_scroll_view"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_weight="1">
-
-        <RelativeLayout
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:orientation="vertical">
-
-            <ImageView
-                android:id="@+id/movieImage"
-                android:layout_width="wrap_content"
-                android:layout_height="100dp"
-                android:layout_alignParentLeft="true"
-                android:layout_alignParentStart="true"
-                android:layout_below="@+id/movieName"
-                android:layout_marginLeft="79dp"
-                android:layout_marginStart="79dp"
-                android:src="@mipmap/ic_launcher" />
-
-            <TextView
-                android:id="@+id/movieName"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:layout_alignParentLeft="true"
-                android:layout_alignParentStart="true"
-                android:layout_alignParentTop="true"
-                android:background="@color/background_color"
-                android:text="@string/movie_name"
-                android:textAppearance="?android:attr/textAppearanceLarge" />
-
-            <Button
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:id="@+id/favorite"
-                android:layout_below="@id/movieName"
-                android:layout_alignParentRight="true"
-                android:text="@string/mark_favorite"
-                />
-
-            <TextView
-                android:id="@+id/synopsis"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_alignParentLeft="true"
-                android:layout_alignParentStart="true"
-                android:layout_below="@+id/movieImage"
-                android:text="@string/synopsis"
-                android:textAppearance="?android:attr/textAppearanceSmall" />
-
-            <TextView
-                android:id="@+id/releaseDate"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_alignEnd="@+id/favorite"
-                android:layout_alignRight="@+id/favorite"
-                android:layout_below="@+id/favorite"
-                android:text="@string/release_date"
-                android:textAppearance="?android:attr/textAppearanceMedium" />
-
-            <TextView
-                android:id="@+id/userRating"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_alignLeft="@+id/releaseDate"
-                android:layout_alignStart="@+id/releaseDate"
-                android:layout_below="@+id/releaseDate"
-                android:text="@string/user_rating"
-                android:textAppearance="?android:attr/textAppearanceSmall" />
-        </RelativeLayout>
-    </ScrollView>
-
-    <ListView
-        android:id="@+id/trailerVideos"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_below="@+id/synopsis"
-        android:layout_weight="1"/>
-
-    <ListView
-        android:id="@+id/userReviews"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_below="@+id/trailerVideos"
-        android:layout_weight="1"/>
-
-</LinearLayout>
+    />
